LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-root-source-expected.txt


FAIL Test view-timeline with document scrolling element. assert_equals: expected "1" but got "0"
